requirements.txt
decorator==4.0.9
Flask==0.10.1
Flask-Cors==2.1.2
Flask-JWT==0.3.2
Flask-Login==0.3.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
infinity==1.4
intervals==0.7.1
itsdangerous==0.24
Jinja2==2.8
ldap3==1.3.1
MarkupSafe==0.23
pbr==1.9.1
psycopg2==2.6.1
pyasn1==0.1.9
PyJWT==1.4.0
python-dateutil==2.5.3
six==1.10.0
SQLAlchemy==1.0.12
sqlalchemy-migrate==0.10.0
SQLAlchemy-Utils==0.32.6
sqlparse==0.1.19
Tempita==0.5.2
validators==0.10.1
Werkzeug==0.11.9
WTForms==2.1
WTForms-Alchemy==0.16.1
WTForms-Components==0.10.0
WTForms-JSON==0.2.10
